tests: Organize tests
