+++ tests/extensions/test_ext_cache.py
